refactor utils into schema and generateDefaultValues
