+++ apps/web/src/components/JSONSchemaForm/Form.tsx
@@ -1,13 +1,13 @@
 import type { JSONSchema7, JSONSchema7Definition } from "json-schema";
 import { FormEvent, type ReactNode } from "react";
-import type { FieldValues, Path, UseFormReturn } from "react-hook-form";
+import type { FieldValues, UseFormReturn } from "react-hook-form";
 import { BooleanField } from "./components/BooleanField.tsx";
 import { JsonTextField } from "./components/JsonTextField.tsx";
 import { NumberField } from "./components/NumberField.tsx";
 import { SelectField } from "./components/SelectField.tsx";
 import { StringField } from "./components/StringField.tsx";
 
-import { formatPropertyName } from "./utils.ts";
+import { formatPropertyName, selectAnyOfSchema } from "./utils/index.ts";
 
 export type SchemaType = string | number | boolean | object | null;
 
@@ -21,122 +21,6 @@ export interface JsonSchemaFormProps<
   // deno-lint-ignore no-explicit-any
   error?: any;
   submitButton: ReactNode;
-}
-
-// Helper function to check if schema type matches runtime type
-function _typeMatches(schemaType: JSONSchema7["type"], runtimeType: string): boolean {
-  if (schemaType === undefined) {
-    return true; // If no type is specified in schema, it matches any type
-  }
-  
-  if (Array.isArray(schemaType)) {
-    return schemaType.some(t => t === runtimeType);
-  }
-  
-  // Special case for integers (typeof returns 'number')
-  if (schemaType === 'integer' && runtimeType === 'number') {
-    return true;
-  }
-  
-  return schemaType === runtimeType;
-}
-
-// Check parent-child relationships for anyOf schema selection
-function _checkParentChildRelationship<T extends FieldValues>(
-  name: string,
-  schema: JSONSchema7,
-  form: JsonSchemaFormProps<T>["form"]
-): JSONSchema7 | null {
-  // Split field name to get parent path
-  const parentPath = name.split('.');
-  if (parentPath.length <= 1) {
-    return null; // No parent exists
-  }
-  
-  // Remove the last segment to get parent path
-  parentPath.pop();
-  const parentName = parentPath.join('.');
-  const parentValue = form.watch(parentName as Path<T>);
-  
-  if (!parentValue || typeof parentValue !== 'object' || parentValue === null) {
-    return null;
-  }
-  
-  // Safely cast parentValue to Record to work with properties
-  const parentRecord = parentValue as Record<string, unknown>;
-  
-  // Try to find schema that best matches based on parent context
-  if (schema.anyOf && Array.isArray(schema.anyOf)) {
-    // First strategy: Look for a type property in parent
-    if ('type' in parentRecord && typeof parentRecord.type === 'string') {
-      // Try to match schema title or a property with parent's type
-      const matchByTitle = schema.anyOf.find(s => {
-        const schemaObj = s as JSONSchema7;
-        return schemaObj.title?.toLowerCase() === parentRecord.type?.toString().toLowerCase();
-      });
-      
-      if (matchByTitle) {
-        return matchByTitle as JSONSchema7;
-      }
-      
-      // Try by specific property that might indicate type
-      const matchByProperty = schema.anyOf.find(s => {
-        const schemaObj = s as JSONSchema7;
-        if (schemaObj.properties) {
-          const props = Object.entries(schemaObj.properties);
-          return props.some(([key, propDef]) => {
-            const propSchema = propDef as JSONSchema7;
-            return (
-              (key === 'type' || key === 'kind' || key.endsWith('Type')) && 
-              propSchema.enum?.includes(parentRecord.type as string)
-            );
-          });
-        }
-        return false;
-      });
-      
-      if (matchByProperty) {
-        return matchByProperty as JSONSchema7;
-      }
-    }
-    
-    // Second strategy: Check if parent has properties that match one schema better
-    return schema.anyOf.find(subSchema => {
-      const schemaObj = subSchema as JSONSchema7;
-      
-      // Skip non-object schemas
-      if (schemaObj.type !== 'object' || !schemaObj.properties) {
-        return false;
-      }
-      
-      // If the schema has a discriminator property, check if it exists in parent
-      // Note: discriminator is an OpenAPI extension not in standard JSON Schema
-      const extendedSchema = schemaObj as JSONSchema7 & {
-        discriminator?: {
-          propertyName?: string;
-          mapping?: Record<string, string>;
-        }
-      };
-      
-      if (extendedSchema.discriminator) {
-        const { propertyName, mapping } = extendedSchema.discriminator;
-        
-        if (propertyName && propertyName in parentRecord) {
-          const discriminatorValue = parentRecord[propertyName];
-          // Check if this value matches any mapping
-          if (mapping && 
-              typeof discriminatorValue === 'string' && 
-              discriminatorValue in mapping) {
-            return true;
-          }
-        }
-      }
-      
-      return false;
-    }) as JSONSchema7 | null;
-  }
-  
-  return null;
 }
 
 export function Form<T extends FieldValues = Record<string, unknown>>(
@@ -216,147 +100,12 @@ function renderField<T extends FieldValues = Record<string, unknown>>(
 ) {
   // Handle anyOf schema
   if (schema.anyOf && Array.isArray(schema.anyOf) && schema.anyOf.length > 0) {
-    // Get the current field value
-    const currentValue = form.watch(name as Path<T>);
-    
-    // If we have a value, try to find a matching schema based on the value's type and structure
-    if (currentValue !== undefined && currentValue !== null) {
-      const valueType = typeof currentValue;
-      
-      // Enhanced type detection for special cases
-      const detectedType = 
-        valueType === 'object' && currentValue === null ? 'null' :
-        valueType === 'object' && Array.isArray(currentValue) ? 'array' :
-        valueType;
-      
-      // First try: find schema with matching type that also validates the data
-      const matchingSchema = schema.anyOf.find(subSchema => {
-        const schemaObj = subSchema as JSONSchema7;
-        
-        // Check if type matches
-        const schemaType = schemaObj.type;
-        if (schemaType === undefined) {
-          return true; // If no type is specified in schema, it matches any type
-        }
-        
-        // Handle array of types
-        if (Array.isArray(schemaType)) {
-          const matchesAny = schemaType.some(t => {
-            if (t === 'integer' && detectedType === 'number') return true;
-            if (t === detectedType) return true;
-            return false;
-          });
-          if (!matchesAny) return false;
-        } 
-        // Handle single type
-        else if (schemaType !== detectedType) {
-          // Special case for integers
-          if (!(schemaType === 'integer' && detectedType === 'number')) {
-            return false;
-          }
-        }
-        
-        // For objects, do additional validation with properties
-        if (detectedType === 'object' && typeof currentValue === 'object' && currentValue !== null && schemaObj.properties) {
-          // Perform the property checking logic here
-          return true; // Simplified for brevity - the detailed implementation is already in the file
-        }
-        
-        // For arrays, perform array validation
-        if (detectedType === 'array' && Array.isArray(currentValue) && currentValue.length > 0 && schemaObj.items) {
-          // Perform the array checking logic here
-          return true; // Simplified for brevity - the detailed implementation is already in the file
-        }
-        
-        return true;
-      });
-      
-      if (matchingSchema) {
-        return renderField<T>(
-          name,
-          matchingSchema as JSONSchema7,
-          form,
-          isRequired,
-          disabled,
-        );
-      }
-    }
-    
-    // If no match based on value, check if any parent fields might help determine the right schema
-    const parentSchema = _checkParentChildRelationship(name, schema, form);
-    if (parentSchema) {
-      return renderField<T>(
-        name,
-        parentSchema,
-        form,
-        isRequired,
-        disabled,
-      );
-    }
-    
-    // Check for child fields that might help determine the right schema
-    const childFields = form.getValues();
-    if (childFields && typeof childFields === 'object') {
-      const namePrefix = `${name}.`;
-      const childFieldPaths = Object.keys(childFields as Record<string, unknown>)
-        .filter(path => path.startsWith(namePrefix));
-      
-      if (childFieldPaths.length > 0) {
-        // Find schema that best matches child field structure
-        const bestMatchSchema = schema.anyOf.find(subSchema => {
-          const schemaObj = subSchema as JSONSchema7;
-          if (schemaObj.type !== 'object' || !schemaObj.properties) {
-            return false;
-          }
-          
-          const schemaProps = Object.keys(schemaObj.properties);
-          
-          // Check if schema properties match child fields
-          for (const childPath of childFieldPaths) {
-            const childName = childPath.slice(namePrefix.length).split('.')[0];
-            if (schemaProps.includes(childName)) {
-              return true;
-            }
-          }
-          
-          return false;
-        });
-        
-        if (bestMatchSchema) {
-          return renderField<T>(
-            name,
-            bestMatchSchema as JSONSchema7,
-            form,
-            isRequired,
-            disabled,
-          );
-        }
-      }
-    }
-    
-    // Default prioritization strategy
-    // First try to find a schema with a default value
-    const schemaWithDefault = schema.anyOf.find(
-      (s) => (s as JSONSchema7).default !== undefined
-    );
-    
-    // Then try to find a non-null schema
-    const nonNullSchema = !schemaWithDefault ? 
-      schema.anyOf.find(s => {
-        const schemaType = (s as JSONSchema7).type;
-        if (Array.isArray(schemaType)) {
-          return !schemaType.includes('null');
-        }
-        return schemaType !== 'null';
-      }) : 
-      null;
-    
-    // Choose the best schema based on priority: default value > non-null > first item
-    const representativeSchema = schemaWithDefault || nonNullSchema || schema.anyOf[0];
+    // Use the unified schema selection utility from schema.ts
+    const representativeSchema = selectAnyOfSchema(schema, undefined, form, name);
     
     return renderField<T>(
       name,
-      representativeSchema as JSONSchema7,
+      representativeSchema,
       form,
       isRequired,
       disabled,
