+++ apps/web/src/components/JSONSchemaForm/index.ts
@@ -1,3 +1,3 @@
 export * from "./Form.tsx";
 export * from "./types.ts";
-export * from "./utils.ts";
+export * from "./utils/index.ts";
