+++ apps/web/src/components/JSONSchemaForm/utils/index.ts
@@ -1,0 +1,4 @@
+// Re-export everything from the individual files
+export * from "./schema.ts";
+export * from "./generateDefaultValues.ts";
+export * from "./generateDefaultValue.ts";
